backend/game_service/requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
bcrypt==4.1.3
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
diff-match-patch==20230430
Django==5.0.6
django-environ==0.11.2
django-import-export==4.1.0
djangorestframework==3.15.2
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
packaging==24.1
pipdeptree==2.23.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
requests==2.32.3
service-identity==24.1.0
setuptools==70.1.0
six==1.16.0
sqlparse==0.5.0
tablib==3.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.2
zope.interface==6.4.post2
